create chance to find crown function that scales exponentially with time played, refactor to new crown directory
