+++ packages/game/src/utils/index.ts
@@ -1,7 +1,6 @@
 import { CommandInteraction } from 'discord.js'
 
 import { asset } from '@adventure-bot/game/utils/asset'
-import { crownArt } from '@adventure-bot/game/utils/crownArt'
 import { d, d6, d20 } from '@adventure-bot/game/utils/dice'
 import { progressBar } from '@adventure-bot/game/utils/progressBar'
 import { randomArrayElement } from '@adventure-bot/game/utils/randomArrayElement'
@@ -9,7 +8,6 @@ import { weightedRandom } from '@adventure-bot/game/utils/weightedRandom'
 import { weightedTable } from '@adventure-bot/game/utils/weightedTable'
 
 export {
-  crownArt,
   randomArrayElement,
   progressBar,
   weightedRandom,
